requirements.txt
fastapi
uvicorn
requests
aiofiles
python-multipart
dotenv
pymupdf
google-genai
google-generativeai
